Aggregate config errors from source-file commands
At the moment it's possible to trample configuration file error reporting by
having an error in fileB which is sourced from fileA at startup.  If there
were any errors in fileA these would be lost because fileB somewhat treads
on the state of fileA.

Because source-file uses the same configuration loading code as the server,
the callers are assumed to have pre-allocated an array of errors to be
filled out as commands are processed.  But each call to source-file will
allocate a new array.  When we end up reporting on this to the user, by
entering copy-mode, this doesn't happen for every allocated buffer, just one
of them.  So not every error ends up being presented to the user.

Instead, make the current cfg_causes array global, have it initialised by
tmux before it parses its main config file, and have source-file also
reference cfg_causes.  This aggregates all the errors across files correctly
and displays them in copy-mode.

Bug originally reported by Sam Livingstone-Gray.