choose-tree: Fix current selection when using '-u'
When choose-tree learnt the '-u; option to render the tree fully expanded,
it didn't then reset the current selection in the expanded tree to be that
of the current session and window.

This introduces a new function window_choose_set_current() which sets the
intended position of the selection when first rendered, and changes
window_choose_ready() to use it.   It's then further used in choose-tree
when '-u' is given to set the selection to the correct position in the
expanded tree.