new-session: Allow start-directory via -c
As with the neww command, this adds a -c option to specify the starting
directory of the session.  This means default-path '-' is honored correctly
without the need to first cd to the directory and start the tmux session from
there.

I personally don't care either way for this patch, but note that one of the
side-effects I'm unsure of is using 'new -c /tmp' would start the initial
window in '/tmp', regardless whether default-path has been set globally.
This might be counter-intuitive.

I did consider changing the way cmd_get_default_path() worked in terms of
not worrying about sessions, but this isn't possible.  Neither have I
bothered defaulting to checking &global_s_options to peturb 'new -c' in case
it's set.

This comes up on IRC a lot.