Give better feedback when setting options
When the 'set' command is used without any switches, the lookup happens
within whichever table is found for the option being set.  Usually this
isn't a problem, since missing -t flags assume a valid context.  However,
when a 'set' command is read from a config file, especially at startup, then
the context is not known and an error is thrown.   But there is never any
message set to explain why the 'set' command didn't work.

This comes up a lot on IRC when people really meant to use '-g' but haven't,
but equally we didn't set an error message either.