configure: fixed to also search setupterm() in libtinfo
ncurses has build time configuration options to separate its different parts
into separate libraries.
If the configure switch '--with-termlib' is provided for example, the
terminfo part gets separtated into a single libtinfo, and thus it is including
setupterm().

This trivial change to configure.ac will change the behavior to also check for
setupterm() in libtinfo. This way the check for setupterm() is not failing on
systems with a separated libtinfo anymore.