notify:  Avoid double-free
Before making notify support hooks, notify_drain() was responsible solely
for free()ing the notify_event.  This was fine, since there was only ever a
single point of entry for it.

But in making notify support hooks, running the commands via
cmdq_hooks_run() now means the free() happens twice:  once from
notify_drain() normally, and the other as a result of finish off running
cmdq_run().

To stop this happening, add a flag to the notify_event structure indicating
when it needs to be free()d and only do it then.