Fix attach without -t crash
The check (in this case) for attach without -t resulted in a segfault,
because tflag was still NULL at the point of looking up the PANE_ID before
the session.

Instead, check for tflag not being NULL before that point, and if it's stll
NULL when looking up the session, don't segfault looking up the other bits
of information for the window, winlink, etc.

Since this could technically apply to sflag as well, also do likewise.