Introduce alerts for hooks
This further builds on alerts with hooks.  A struct alert is created
whenever an alert is found to have a hook.  In it, are the session and the
winlinks for that alert.

When the hook command is run, the state is checked in the usual way, but
this time, when looking up the session, window and pane, preference is given
to those items having an alert.

This allows hooked commands on bell events to fire on the correct session.
Since alerts operate outside of the usual way commands make it to tmux (they
can happen at any time), this seems like a sane approach.

Making alerts an inherent property of the session might be an idea for the
future.