Window flags: window_linked
This primarily adds support for a new flag, 'window_linked' to indicate
whether the given window is linked across multiple sessions.  In the case of
grouped sessions, only those windows explicitly linked via lastw are
counted.  This is so that the flag doesn't always return 1 for windows in
grouped sessions.

To make this flag printable as part of the window_flags without hardcoding
it, and to not leave an extra space present (because no window flags means a
single space is printed), this extra space has been removed, and the default
options changed to add the space back in, so there's no regressions.

This means it's now entirely possible to override the window printable flags
through formats alone, albeit long-winded.